moving functions to a helper (instead of piggying backing on types)
