+++ src/components/types.ts
@@ -292,30 +292,3 @@ export interface GridItem {
     type1: string,
     type2: string,
 } 
-
-//functions
-
-export function getStatLabel(statName: string){
-    let output = '';
-    switch (statName) {
-      case 'hp':
-        output = statName; 
-        break;
-      case 'attack':
-        output = 'atk';
-        break;
-      case 'defense':
-        output = 'def';
-        break;
-      case 'special-attack':
-        output = 'spa';
-        break;
-      case 'special-defense':
-        output = 'spd';
-        break;
-      case 'speed':
-        output = 'spe';
-        break;
-    }
-    return output.toUpperCase();
-}
